@@ -79,17 +79,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/schematics": "^13.2.6",
-    "@angular/animations": "12.2.6",
-    "@angular/common": "12.2.6",
-    "@angular/compiler": "12.2.6",
-    "@angular/core": "12.2.6",
-    "@angular/forms": "12.2.6",
-    "@angular/platform-browser": "12.2.6",
-    "@angular/platform-browser-dynamic": "12.2.6",
-    "@angular/platform-server": "12.2.6",
-    "@angular/router": "12.2.6",
-    "@nrwl/nx-cloud": "^12.3.13",
+    "@angular-devkit/schematics": "14.2.3",
+    "@angular/animations": "14.2.3",
+    "@angular/common": "14.2.3",
+    "@angular/compiler": "14.2.3",
+    "@angular/core": "14.2.3",
+    "@angular/forms": "14.2.3",
+    "@angular/platform-browser": "14.2.3",
+    "@angular/platform-browser-dynamic": "14.2.3",
+    "@angular/platform-server": "14.2.3",
+    "@angular/router": "14.2.3",
+    "@nrwl/nx-cloud": "14.6.2",
     "@sentry/node": "^6.12.0",
     "@types/compression": "^1.7.2",
     "@types/cors": "^2.8.12",
@@ -174,30 +173,36 @@
     "guess-parser": "^0.4.22",
     "http-proxy-middleware": "^2.0.3",
     "husky": "^7.0.4",
-    "jasmine-core": "3.9.0",
+    "jasmine-core": "4.2.0",
     "jasmine-spec-reporter": "7.0.0",
-    "jest": "27.2.0",
-    "jest-preset-angular": "11.1.1",
+    "jest": "28.1.3",
+    "jest-environment-jsdom": "28.1.1",
+    "jest-preset-angular": "12.2.2",
     "jest-serializer-html": "^7.1.0",
     "jsdom": "^19.0.0",
     "jsonc": "^2.0.0",
     "marked": "^4.0.12",
-    "ng-packagr": "12.2.1",
+    "ng-packagr": "14.2.1",
     "nodemon": "^2.0.15",
+    "nx": "14.7.16",
     "open": "^8.4.0",
     "osascript": "^1.2.0",
     "path-to-regexp": "^6.2.0",
-    "prettier": "2.5.1",
+    "postcss": "^8.3.9",
+    "postcss-import": "^14.0.2",
+    "postcss-preset-env": "^6.7.0",
+    "postcss-url": "^10.1.1",
+    "prettier": "2.7.1",
     "pretty-quick": "^3.1.3",
     "protractor": "7.0.0",
     "simple-git": "^3.3.0",
     "start-server-and-test": "^1.14.0",
     "symlink-dir": "^5.0.1",
-    "ts-jest": "27.1.3",
-    "ts-node": "^10.5.0",
+    "ts-jest": "28.0.8",
+    "ts-node": "10.9.1",
     "tsickle": "^0.46.0",
     "tslib": "^2.3.1",
-    "typescript": "4.3.5",
+    "typescript": "4.7.4",
     "wait-on": "^6.0.0",
     "yamljs": "^0.3.0",
     "yargs": "^17.3.1"
