@@ -1,6 +1,7 @@
-module.exports = {
-  displayName: 'scully-schematics',
-  preset: '../../jest.preset.js',
+/* eslint-disable */
+export default {
+  displayName: 'plugins-scully-plugin-puppeteer',
+  preset: '../../../jest.preset.js',
   globals: {
     'ts-jest': {
       tsconfig: '<rootDir>/tsconfig.spec.json',
@@ -11,6 +12,5 @@ module.exports = {
     '^.+\\.[tj]sx?$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../coverage/libs/scully-schematics',
-  testPathIgnorePatterns: ['<rootDir>/src/files/'],
+  coverageDirectory: '../../../coverage/libs/plugins/scully-plugin-puppeteer',
 };
