@@ -1,5 +1,5 @@
-const nxPreset = require('@nrwl/jest/preset');
+const nxPreset = require('@nrwl/jest/preset').default;
 
 module.exports = {
-  ...nxPreset
+  ...nxPreset,
 };
